// idea-investment-webapp-main/src/enums/notification.enum.ts
// export enum NOTIFICATION_MODIFIABLE {
//   seen = 'seen',
//   read = 'read',
//   sent = 'sent',
// }
export enum NOTIFICATION_STATUS {
  publish = 'publish',
  save = 'save',
  unpublish = 'unpublish',
}
export enum DEVICE {
  all = 'all',
  ios = 'ios',
  android = 'android',
}
export enum NOTIFICATIONTYPE {
  Orderid = 'ORDERID',
  Coupon = 'COUPON',
  Project = 'PROJECT',
  Offer = 'OFFER',
  Link = 'LINK',
  ScreenName = 'SCREENNAME',
  URL = 'URL',
}